+++ packages/rb-windowing/src/index.ts
@@ -6,3 +6,4 @@ export * from './types';
 export * from './store';
 export * from './focusTrap';
 export * from './invariants';
+export * from './windowRouting';